+passed autotests iter1-8 with new architecture
