+++ internal/entity/entity.go
@@ -1,6 +1,11 @@
-// Package entity defines main entities for business logic (services), data base mapping and
+// Package entity defines main entities for business logic (services), database mapping and
 // HTTP response objects if suitable. Each logic group entities in own file.
 package entity
+
+var URLStorage Keeper
+var HostSocketAddrStr string
+var BaseURLStr string
+var RepoFilePathStr string
 
 type TxJSONMessage struct {
 	URL string `json:"url"`
